@@ -42,13 +42,10 @@ static const char* STR_READ_LENGTH = "-read-length";
 static const char* STR_MEAN = "-mean";
 static const char* STR_STD_DEV = "-std-dev";
 
-static const char* STR_SCAFFOLD = "-scaffold";
-static const char* STR_GAP_BREAKPOINT = "-breakpoint";
+static const char* STR_REGION = "-region";
 
-static const char* STR_FLANK_LENGTH = "-flank-length";
-
-static const char* STR_GAP_LENGTH = "-gap-length";
 static const char* STR_THRESHOLD = "-unmapped";
+static const char* STR_ONLY_UNMAPPED= "-unmapped-only";
 
 /*****************************************************************************/
 
@@ -63,13 +60,11 @@ Extract::Extract() : Tool("Extract") {
   getParser()->push_front(new OptionOneParam(STR_STD_DEV, "Insert size standard deviation", true));
 
   // Gap parameters
-  getParser()->push_front(new OptionOneParam(STR_SCAFFOLD, "Scaffold name", true));
-  getParser()->push_front(new OptionOneParam(STR_GAP_BREAKPOINT, "Gap position", true));
+  getParser()->push_front(new OptionOneParam(STR_REGION, "Region for filtering", true));
 
-  getParser()->push_front(new OptionOneParam(STR_FLANK_LENGTH , "Flank length", false, "-1"));
-
-  getParser()->push_front(new OptionOneParam(STR_GAP_LENGTH, "Gap length", false, "-1"));
-  getParser()->push_front(new OptionOneParam(STR_THRESHOLD, "Threshold for using unmapped reads", false, "-1"));
+  // Unmapped read parameters
+  getParser()->push_front(new OptionOneParam(STR_THRESHOLD, "Threshold for using unmapped reads", false, "0"));
+  getParser()->push_front(new OptionNoParam(STR_ONLY_UNMAPPED, "Only output unmapped reads"));
 }
 
 /*****************************************************************************/
@@ -166,59 +161,69 @@ void Extract::execute() {
   const std::string alignment = getInput()->getStr(STR_ALIGNMENT);
   const std::string output = getInput()->getStr(STR_OUTPUT);
 
+  // TODO: Infer read length (assume all reads are same length or nah?)
   const int read_length = static_cast<int>(getInput()->getInt(STR_READ_LENGTH));
   const int mean_insert = static_cast<int>(getInput()->getInt(STR_MEAN));
   const int std_dev = static_cast<int>(getInput()->getInt(STR_STD_DEV));
 
-  const std::string scaffold = getInput()->getStr(STR_SCAFFOLD);
-  const int breakpoint = static_cast<int>(getInput()->getInt(STR_GAP_BREAKPOINT));
+  const std::string region = getInput()->getStr(STR_REGION);
 
-  const int flank_length = static_cast<int>(getInput()->getInt(STR_FLANK_LENGTH));
-
-  const int gap_length = static_cast<int>(getInput()->getInt(STR_GAP_LENGTH));
   const int threshold = static_cast<int>(getInput()->getInt(STR_THRESHOLD));
+  const bool unmapped_only = getParser()->saw(STR_ONLY_UNMAPPED);
+
+  // Parse samtools style region
+  const size_t comma = region.find(":");
+  const size_t dash = region.find("-");
+  if (comma == std::string::npos || dash == std::string::npos) {
+    std::cerr << "Malformed region, should be CONTIG:START-END" << std::endl;
+    return;
+  }
+
+  const std::string scaffold = region.substr(0, comma);
+  const int start = std::stoi(region.substr(comma+1, dash-comma));
+  const int end = std::stoi(region.substr(dash+1));
 
   // Load alignment file
   io_t io(alignment);
-  if (!io.loaded) {
+  if (!io.loaded()) {
     std::cerr << "Error loading alignments" << std::endl;
     return;
   }
 
-  // Compute scaffold id from scaffold name
-  const int tid = bam_name2id(io.header, scaffold.c_str());
+  // Allocate memory for string conversions
+  char *buffer = new char[read_length+1];
 
   // Use basic Bloom filter from GATB
   const uint64_t num_of_reads = count_reads(alignment);
   IBloom<std::string> *bloom = new BloomSynchronized<std::string>(5 * num_of_reads);
 
-  // Extract pairs from the left mappings
-  const int left_start = breakpoint - (mean_insert + 3*std_dev + 2*read_length);
-  const int left_end = breakpoint - (mean_insert - 3*std_dev + read_length);
-  process_mates(io, tid, left_start, left_end, bloom);
-
-  // Extract pairs from the right mappings
-  const int right_start = breakpoint + (mean_insert + 3*std_dev + read_length);
-  const int right_end = breakpoint + (mean_insert - 3*std_dev + read_length);
-  process_mates(io, tid, right_start, right_end, bloom);
-
-  // Allocate memory for string conversions
-  char *buffer = new char[read_length+1];
-
-  // Output reads and count length
+  // Open output file
   BankFasta reads(output);
   int seqlen = 0, reads_extracted = 0;
-  find_mates(io, buffer, bloom, &reads, &seqlen, &reads_extracted);
 
-  // Output overlapping reads
-  if (flank_length != -1) {
-    const int start = breakpoint - flank_length;
-    const int end = breakpoint + flank_length;
+  if (!unmapped_only) {
+    // Compute scaffold id from scaffold name
+    const int tid = bam_name2id(io.header, scaffold.c_str());
+
+    // Extract pairs from the left mappings
+    const int left_start = start - (mean_insert + 3*std_dev + 2*read_length);
+    const int left_end = end - (mean_insert - 3*std_dev + read_length);
+    process_mates(io, tid, left_start, left_end, bloom);
+
+    // Extract pairs from the right mappings
+    const int right_start = start + (mean_insert + 3*std_dev + read_length);
+    const int right_end = end + (mean_insert - 3*std_dev + read_length);
+    process_mates(io, tid, right_start, right_end, bloom);
+
+    // Output reads and count length
+    find_mates(io, buffer, bloom, &reads, &seqlen, &reads_extracted);
+
+    // Output overlapping reads
     process_region(io, tid, start, end, buffer, bloom, &reads, &seqlen, &reads_extracted);
   }
 
   // Output unmapped reads
-  if ((gap_length != -1 && threshold != -1) && ((seqlen / gap_length) < threshold)) {
+  if (unmapped_only || (seqlen / (end - start)) < threshold) {
     process_unmapped(io, buffer, bloom, &reads, &reads_extracted);
   }
 
